Refactor scroll to top as reusable component
