+++ client/app/components/authors/Show.jsx
@@ -1,15 +1,15 @@
-import React, { useState, useEffect } from "react";
+import React, { useState } from "react";
 import axios from "axios";
 import SVG from "react-inlinesvg";
 import Post from "../posts/Post";
+import ScrollToTopButton from "../shared/ScrollToTopButton";
 import getAuthToken from "../../utils/getAuthToken";
-import { IcChevronDown, IcArrowLong } from "../../assets/icons";
+import { IcChevronDown } from "../../assets/icons";
 import "./Show.scss";
 
 const Show = ({ posts, olderThan, displayAuthor }) => {
     const [visiblePosts, setVisiblePosts] = useState(posts);
     const [loadMorePostsDate, setLoadMorePostsDate] = useState(olderThan);
-    const [showScrollToTop, setShowScrollToTop] = useState(false);
 
     const loadMorePosts = () => {
         axios
@@ -28,20 +28,6 @@ const Show = ({ posts, olderThan, displayAuthor }) => {
                 setLoadMorePostsDate(null);
             })
     };
-
-    const updateScrollToTop = () => {
-        const scrollHeight = window.pageYOffset;
-        setShowScrollToTop(scrollHeight > 500);
-    };
-
-    const scrollToTop = () => {
-        window.scrollTo({ top: 0, left: 0, behavior: "smooth"});
-    }
-    
-    useEffect(() => {
-        window.addEventListener("scroll", updateScrollToTop);
-        return () => window.removeEventListener("scroll", updateScrollToTop);
-    }, []);
 
     return (
         <div id="author-profile">
@@ -65,17 +51,7 @@ const Show = ({ posts, olderThan, displayAuthor }) => {
                     </div>
                 )}
             </div>
-            <button
-                className={`button scroll-to-top__button ${showScrollToTop ? "scroll-to-top__button--visible" : ""}`}
-                onClick={scrollToTop}
-            >
-                <div className="scroll-to-top__container">
-                    <SVG
-                        src={IcArrowLong}
-                        className="scroll-to-top__icon"
-                    />
-                </div>
-            </button>
+            <ScrollToTopButton />
         </div>
     );
 };
