+++ client/app/components/authors/Show.scss
@@ -41,48 +41,3 @@
         padding: $spacing-56 $spacing-136;
     }
 }
-
-.button.scroll-to-top__button {
-    background-color: $color-white;
-    position: fixed;
-    bottom: $spacing-16;
-    right: $spacing-16;
-    padding: 0;
-    border-radius: 22px;
-    display: none;
-
-    .scroll-to-top__container {
-        width: 44px;
-        height: 44px;
-        border: $border-default;
-        border-radius: 22px;
-        background-color: $color-default;
-        display: flex;
-        align-items: center;
-        justify-content: center;
-        opacity: 0;
-        transition: opacity 0.1s ease-in-out;
-    
-        .scroll-to-top__icon {
-            width: 24px;
-            height: 24px;
-            transform: rotate(-90deg);
-    
-            path {
-                fill: $color-black;
-            }
-        }
-
-        &:hover {
-            border: 1px solid $color-black;
-        }
-    }
-}
-
-.button.scroll-to-top__button--visible {
-    display: block;
-
-    .scroll-to-top__container {
-        opacity: 1;
-    }
-}
